Travel/flight_booking.py
import requests
from bs4 import BeautifulSoup
from datetime import datetime

# Get event details from the user
departure = input("Enter the name of the event: ")
arrival = input("Enter the city where the event is taking place: ")
travel_date = input("Enter the date of the event (MM/DD/YYYY): ")
num_tickets = int(input("Enter the number of tickets you want to purchase: "))

# Get user's personal information
first_name = input("Enter your first name: ")
last_name = input("Enter your last name: ")
email = input("Enter your email address: ")
phone = input("Enter your phone number: ")

# Define the payment information (replace with your own card details)
card_number = "1234567890123456"
expiry_date = "12/25"
cvv = "123"

# Convert the event date to the required format
event_date_formatted = datetime.strptime(event_date, "%m/%d/%Y").strftime("%d/%m/%Y")

# Construct the URL for the event search
search_url = f"https://www.makemytrip.com/flight/search?itinerary={departure}-{arrival}-{travel_date}"

# Navigate to the event search page
session = requests.Session()
response = session.get(search_url)
soup = BeautifulSoup(response.content, "html.parser")

# Find the event listing and extract the event URL
event_listing = soup.find("div", class_="event-listing")
if event_listing:
    event_url = "https://www.makemytrip.com" + event_listing.find("a")["href"]
else:
    print("Could not find the event. Please check the details and try again.")
    exit()


# Find the "Book Tickets" button and click it
book_button = soup.find("button", class_="book-btn")
if book_button:
    book_url = "https://www.makemytrip.com" + book_button["data-url"]
    response = session.get(book_url)
    soup = BeautifulSoup(response.content, "html.parser")
else:
    print("Could not find the 'Book Tickets' button. Please try again later.")
    exit()

# Fill out the form with user's personal and payment information
# (You'll need to inspect the HTML structure of the booking page to find the appropriate form fields)
# ...

# Submit the form to purchase the tickets
# ...

# Check if the purchase was successful
# ...